Separated out rivers and bridges into its own file with three map variants
